[x86][pci] Introduce the PCI RootHost concept
PciRootHost replaces the static allocators used in the
x86 board file with an interface that can be implemented
on different platforms for PCI Root support. Additionally,
Pciroot is now templated on the type of platform context necessary,
so bringing up PCI on a non-ACPI system should be similar to setting
up the basic board file as one would for a normal driver. In subsequent
CLs RootHost will be expanded to handle out of process resource
management centrally.

Some other changes were made to make the resource logic in the ACPI
parsing more clear to folks who don't speak ACPICA.

Change-Id: I5994b056e39f478653d09cef1f589c0f80bcc206